refactor(core): reorganize tests and actions structure
