+++ packages/core/src/actions/callback/callback.ts
@@ -4,13 +4,12 @@ import { createCSRF } from "@/shared/crypto.ts"
 import { parse } from "@aura-stack/router/cookie"
 import { cacheControl } from "@/shared/headers.ts"
 import { AuraAuthError } from "@/shared/errors.ts"
-import { timingSafeEqual, transformToTokenPayload } from "@/shared/utils.ts"
 import { getCookie, getExpiredCookie } from "@/cookie.ts"
-import { getUserInfo } from "@/actions/callback/userinfo.ts"
 import { validateIDToken } from "@/shared/oidc/id-token.ts"
-import { createAccessToken } from "@/actions/callback/access-token.ts"
+import { createAccessToken, getUserInfo } from "@/shared/utils/oauth.ts"
+import { timingSafeEqual, transformToTokenPayload } from "@/shared/utils.ts"
 import { isRelativeURL, isSameOrigin, isTrustedOrigin } from "@/shared/assert.ts"
-import { getOriginURL, getTrustedOrigins } from "@/actions/signIn/authorization.ts"
+import { getOriginURL, getTrustedOrigins } from "@/shared/utils/authorization.ts"
 import { isOIDCProvider, resolveOpenIDProvider } from "@/shared/oidc/resolve-provider.ts"
 import type { OAuthProviderRecord } from "@/@types/index.ts"
