+++ packages/core/src/shared/utils/authorization.ts
@@ -1,9 +1,11 @@
 import { getEnv } from "@/shared/env.ts"
+import { OAuthAuthorization } from "@/schemas.ts"
 import { AuraAuthError } from "@/shared/errors.ts"
+import { createPKCE, createSecretValue } from "@/shared/crypto.ts"
 import { equals, extractPath, patternToRegex } from "@/shared/utils.ts"
 import { isRelativeURL, isSameOrigin, isValidURL, isTrustedOrigin } from "@/shared/assert.ts"
-import type { AuthConfig } from "@/@types/index.ts"
 import type { GlobalContext } from "@aura-stack/router"
+import type { AuthConfig, OAuthProvider } from "@/@types/index.ts"
 import type { Identities, SchemaTypes } from "@/identity/index.ts"
 
 /**
@@ -138,5 +140,86 @@ export const createRedirectTo = async (request: Request, redirectTo?: string, co
     } catch {
         context?.logger?.log("POTENTIAL_OPEN_REDIRECT_ATTACK_DETECTED")
         return "/"
+    }
+}
+
+export const setSearchParams = (url: URL, params: Record<string, string | undefined>) => {
+    for (const [key, value] of Object.entries(params)) {
+        if (value !== undefined && value !== "") {
+            url.searchParams.set(key, value)
+        }
+    }
+}
+
+export const buildAuthorizationURL = (
+    oauth: OAuthProvider,
+    redirect_uri: string,
+    state: string,
+    code_challenge: string,
+    code_challenge_method: string
+): string => {
+    const authorizeConfig = oauth.authorize
+    const baseURL = typeof authorizeConfig === "string" ? authorizeConfig : (authorizeConfig?.url ?? oauth.authorizeURL)
+    if (!baseURL) {
+        throw new AuraAuthError({ code: "INVALID_OAUTH_PROVIDER_URL_CONFIG" })
+    }
+    let url: URL
+    try {
+        url = new URL(baseURL)
+    } catch (cause) {
+        throw new AuraAuthError({ code: "INVALID_OAUTH_PROVIDER_URL_CONFIG", cause })
+    }
+    const authorizeParams = typeof authorizeConfig === "string" ? undefined : authorizeConfig?.params
+
+    setSearchParams(url, {
+        response_type: authorizeParams?.responseType ?? oauth.responseType ?? "code",
+        client_id: oauth.clientId,
+        redirect_uri,
+        state,
+        code_challenge,
+        code_challenge_method,
+        scope: authorizeParams?.scope ?? oauth.scope,
+        prompt: authorizeParams?.prompt,
+        response_mode: authorizeParams?.responseMode,
+        login_hint: authorizeParams?.loginHint,
+        nonce: authorizeParams?.nonce,
+        display: authorizeParams?.display,
+        audience: authorizeParams?.audience,
+    })
+    return url.toString()
+}
+
+/**
+ * Constructs the request URI for the Authorization Request to the third-party OAuth service. It includes
+ * the necessary query parameters such as `client_id`, `redirect_uri`, `response_type`, `scope`, `state`,
+ * `code_challenge`, and `code_challenge_method`.
+ *
+ * @see https://datatracker.ietf.org/doc/html/rfc6749#section-4.1.1
+ * @see https://datatracker.ietf.org/doc/html/rfc7636#section-4
+ */
+export const createAuthorizationURL = async (oauth: OAuthProvider, redirectURI: string, ctx?: GlobalContext) => {
+    const state = createSecretValue()
+    const { codeVerifier, codeChallenge, method } = await createPKCE()
+    const authorization = buildAuthorizationURL(oauth, redirectURI, state, codeChallenge, method)
+
+    const parsed = OAuthAuthorization.safeParse({ ...oauth, redirectURI, state, codeChallenge, codeChallengeMethod: method })
+    if (!parsed.success) {
+        ctx?.logger?.log("INVALID_OAUTH_CONFIGURATION", {
+            structuredData: {
+                scope: oauth?.scope ?? "",
+                redirect_uri: redirectURI,
+                has_state: Boolean(state),
+                has_code_challenge: Boolean(codeChallenge),
+                code_challenge_method: method,
+            },
+        })
+        throw new AuraAuthError({ code: "INVALID_OAUTH_PROVIDER_SCHEMA_CONFIG" })
+    }
+
+    return {
+        authorization,
+        state,
+        codeVerifier,
+        method,
     }
 }
