--- packages/core/test/actions/signIn/authorization.test.ts
@@ -1,863 +1,0 @@
-import { afterEach, beforeEach, describe, expect, test, vi } from "vitest"
-import { oauthCustomService } from "@test/presets.ts"
-import { createAuthorizationURL } from "@/actions/signIn/authorization-url.ts"
-import { createRedirectTo, createRedirectURI, getBaseURL, getOriginURL } from "@/actions/signIn/authorization.ts"
-import type { OAuthProviderCredentials } from "@/@types/index.ts"
-import type { GlobalContext } from "@aura-stack/router/types"
-
-beforeEach(() => {
-    vi.stubEnv("BASE_URL", undefined)
-})
-
-afterEach(() => {
-    vi.unstubAllEnvs()
-})
-
-describe("createRedirectURI", () => {
-    describe("valid request", () => {
-        const testCases = [
-            {
-                description: "creates redirect URI from standard URL",
-                request: new Request("https://example.com/signIn/github"),
-                oauth: "github",
-                expected: "https://example.com/auth/callback/github",
-                trustedProxyHeaders: false,
-            },
-            {
-                description: "creates redirect URI from URL with port",
-                request: new Request("http://localhost:3000/signIn/google"),
-                oauth: "google",
-                expected: "http://localhost:3000/auth/callback/google",
-                trustedProxyHeaders: false,
-            },
-            {
-                description: "creates redirect URI from URL with path and query",
-                request: new Request("https://example.com/signIn/github?query=123"),
-                oauth: "facebook",
-                expected: "https://example.com/auth/callback/facebook",
-                trustedProxyHeaders: false,
-            },
-            {
-                description: "creates redirect URI from URL with trailing slash",
-                request: new Request("https://example.com/signIn/github/"),
-                oauth: "github",
-                expected: "https://example.com/auth/callback/github",
-                trustedProxyHeaders: false,
-            },
-            {
-                description: "creates redirect URI from URL with undefined query",
-                request: new Request("https://example.com/signIn/github?"),
-                oauth: "github",
-                expected: "https://example.com/auth/callback/github",
-                trustedProxyHeaders: false,
-            },
-            {
-                description: "creates redirect URI from URL with hash",
-                request: new Request("https://example.com/signIn/twitter#section"),
-                oauth: "twitter",
-                expected: "https://example.com/auth/callback/twitter",
-                trustedProxyHeaders: false,
-            },
-            {
-                description: "creates redirect URI from URL with subdomain",
-                request: new Request("https://subdomain.example.com/signIn/linkedin"),
-                oauth: "linkedin",
-                expected: "https://subdomain.example.com/auth/callback/linkedin",
-                trustedProxyHeaders: false,
-            },
-            {
-                description: "creates redirect URI from URL with IP address",
-                request: new Request("http://192.168.1.1/signIn/github"),
-                oauth: "github",
-                expected: "http://192.168.1.1/auth/callback/github",
-                trustedProxyHeaders: false,
-            },
-        ]
-
-        for (const { description, request, oauth, expected, trustedProxyHeaders } of testCases) {
-            test(description, async () => {
-                const redirectURI = await createRedirectURI(request, oauth, {
-                    basePath: "/auth",
-                    trustedProxyHeaders,
-                } as GlobalContext)
-                expect(redirectURI).toBe(expected)
-            })
-        }
-    })
-
-    describe("invalid request", () => {
-        const testCases = [
-            {
-                description: "creates redirect URI from URL with trusted proxy headers",
-                request: new Request("http://localhost:3000/signIn/google", {
-                    headers: {
-                        "X-Forwarded-Proto": "https",
-                        "X-Forwarded-Host": "example.com",
-                    },
-                }),
-                oauth: "google",
-                trustedProxyHeaders: true,
-            },
-        ]
-
-        for (const { description, request, oauth, trustedProxyHeaders } of testCases) {
-            test(description, async () => {
-                await expect(
-                    createRedirectURI(request, oauth, {
-                        basePath: "/auth",
-                        trustedProxyHeaders,
-                    } as GlobalContext)
-                ).rejects.toThrow(
-                    "The request location context was blocked. The incoming value does not match patterns mapped inside your array configuration rules."
-                )
-            })
-        }
-    })
-})
-
-describe("createAuthorizationURL", () => {
-    vi.stubEnv("BASE_URL", undefined)
-
-    describe("valid OAuth configuration", () => {
-        test("valid OAuth configuration with all fields", async () => {
-            const { authorization } = await createAuthorizationURL(
-                oauthCustomService,
-                "https://example.com/auth/callback/oauth-provider"
-            )
-
-            const searchParams = new URL(authorization).searchParams
-            expect(searchParams.has("client_secret")).toBeFalsy()
-            expect(searchParams.get("client_id")).toBe("oauth_client_id")
-            expect(searchParams.get("redirect_uri")).toBe("https://example.com/auth/callback/oauth-provider")
-            expect(searchParams.get("state")).toBeTruthy()
-            expect(searchParams.get("scope")).toBe("profile email")
-            expect(searchParams.get("response_type")).toBe("code")
-        })
-    })
-
-    describe("invalid OAuth configuration", () => {
-        const testCases = [
-            {
-                description: "missing clientId",
-                oauthConfig: {
-                    authorizeURL: "https://example.com/oauth/authorize",
-                    accessToken: "https://example.com/oauth/access_token",
-                    userInfo: "https://example.com/oauth/user_info",
-                    scope: "read:user",
-                    responseType: "code",
-                    clientSecret: "2",
-                },
-                redirectURL: "https://example.com/auth/callback",
-                expected:
-                    "The loaded configuration settings array failed standard library schema validation checks against required engine operational footprints.",
-            },
-            {
-                description: "missing clientSecret",
-                oauthConfig: {
-                    authorizeURL: "https://example.com/oauth/authorize",
-                    accessToken: "https://example.com/oauth/access_token",
-                    userInfo: "https://example.com/oauth/user_info",
-                    scope: "read:user",
-                    responseType: "code",
-                    clientId: "1",
-                },
-                redirectURL: "https://example.com/auth/callback",
-                expected:
-                    "The loaded configuration settings array failed standard library schema validation checks against required engine operational footprints.",
-            },
-        ]
-
-        for (const { description, oauthConfig, redirectURL, expected } of testCases) {
-            test(description, async () => {
-                await expect(() =>
-                    createAuthorizationURL(oauthConfig as unknown as OAuthProviderCredentials, redirectURL, {} as GlobalContext)
-                ).rejects.toThrow(expected)
-            })
-        }
-    })
-})
-
-describe("createRedirectTo", () => {
-    const signInURL = "https://example.com/auth/signIn/github"
-
-    describe("valid origin", () => {
-        const testCases = [
-            {
-                description: "returns '/' when Referer header is missing",
-                request: new Request(signInURL),
-                expected: "/",
-            },
-            {
-                description: "returns pathname when origins match",
-                request: new Request(signInURL, {
-                    headers: { Referer: "https://example.com/auth" },
-                }),
-                expected: "/auth",
-            },
-            {
-                description: "returns pathname with trailing slash when origins match",
-                request: new Request(signInURL, {
-                    headers: { Referer: "https://example.com/auth/" },
-                }),
-                expected: "/auth/",
-            },
-
-            {
-                description: "pathname contains multiple consecutive slashes",
-                request: new Request(signInURL, {
-                    headers: { Referer: "https://example.com//auth///dashboard" },
-                }),
-                expected: "/",
-            },
-            {
-                description: "pathname contains invalid segments (parent directory)",
-                request: new Request(signInURL, {
-                    headers: { Referer: "https://example.com/../admin" },
-                }),
-                expected: "/",
-            },
-            {
-                description: "pathname is root '/' with referer header",
-                request: new Request(signInURL, {
-                    headers: { Referer: "https://example.com" },
-                }),
-                expected: "/",
-            },
-            {
-                description: "with jump lines in Referer header",
-                request: new Request(signInURL, {
-                    headers: { Referer: "https://example.com/auth/\r\n" },
-                }),
-                expected: "/auth/",
-            },
-            {
-                description: "with localhost in Referer header",
-                request: new Request("http://localhost:3000/auth/signIn/github", {
-                    headers: { Referer: "http://localhost:3000/dashboard" },
-                }),
-                expected: "/dashboard",
-            },
-            {
-                description: "with underscores in Referer header",
-                request: new Request("https://example_name.com/auth/signIn/github", {
-                    headers: { Referer: "https://example_name.com/auth" },
-                }),
-                expected: "/auth",
-            },
-            {
-                description: "with complex path in Referer header",
-                request: new Request("https://example_name.com/auth/signIn/github", {
-                    headers: { Referer: "https://example_name.com/dashboard/admin/panel" },
-                }),
-                expected: "/dashboard/admin/panel",
-            },
-            {
-                description: "with origin header matching hosted origin",
-                request: new Request(signInURL, {
-                    headers: { Origin: "https://example.com" },
-                }),
-                expected: "/",
-            },
-            {
-                description: "with origin and referer headers matching hosted origin",
-                request: new Request(signInURL, {
-                    headers: {
-                        Origin: "https://example.com",
-                        Referer: "https://example.com/some/path",
-                    },
-                }),
-                expected: "/some/path",
-            },
-            {
-                description: "pathname is '/' with origin header",
-                request: new Request(signInURL, {
-                    headers: { Origin: "https://example.com" },
-                }),
-                expected: "/",
-            },
-            {
-                description: "with path in origin header",
-                request: new Request(signInURL, {
-                    headers: { Origin: "https://example.com/dashboard" },
-                }),
-                expected: "/dashboard",
-            },
-            {
-                description: "with redirectTo parameter provided",
-                request: new Request(signInURL),
-                redirectTo: "/auth/signIn",
-                expected: "/auth/signIn",
-            },
-            {
-                description: "with redirectTo parameter matching hosted origin",
-                request: new Request(signInURL),
-                redirectTo: "https://example.com/auth/signIn",
-                expected: "/auth/signIn",
-            },
-            {
-                description: "with redirectTo parameter overriding referer header",
-                request: new Request(signInURL, {
-                    headers: { Referer: "https://example.com/dashboard" },
-                }),
-                redirectTo: "/auth/signIn",
-                expected: "/auth/signIn",
-            },
-            {
-                description: "with localhost and without redirectTo parameter",
-                request: new Request("http://localhost:3000/auth/signIn/github"),
-                redirectTo: "/",
-                expected: "/",
-            },
-            {
-                description: "with localhost and referer header",
-                request: new Request("http://localhost:3000/auth/signIn/github", {
-                    headers: { Referer: "http://localhost:3000/dashboard" },
-                }),
-                redirectTo: "/dashboard",
-                expected: "/dashboard",
-            },
-            {
-                description: "with Ipv4 address and without referer header",
-                request: new Request("http://192.168.0.1/auth/signIn/github"),
-                redirectTo: "/",
-                expected: "/",
-            },
-            {
-                description: "with Ipv4 address and referer header",
-                request: new Request("http://192.168.0.1/auth/signIn/github", {
-                    headers: { Referer: "http://192.168.0.1/dashboard" },
-                }),
-                redirectTo: "/dashboard",
-                expected: "/dashboard",
-            },
-            {
-                description: "with IPv4 address and referer header",
-                request: new Request("https://192.168.0.1/auth/signIn/github", {
-                    headers: { Referer: "https://192.168.0.1/dashboard" },
-                }),
-                redirectTo: "/dashboard",
-                expected: "/dashboard",
-            },
-            {
-                description: "with trusted proxy headers enabled",
-                request: new Request("http:/localhost:3000/auth/signIn/github", {
-                    headers: {
-                        "X-Forwarded-Proto": "https",
-                        "X-Forwarded-Host": "http://192.168.0.1",
-                    },
-                }),
-                trustedProxyHeaders: true,
-                redirectTo: "/",
-                expected: "/",
-            },
-            {
-                description: "with trusted proxy headers disabled",
-                request: new Request("http:/localhost:3000/auth/signIn/github", {
-                    headers: {
-                        "X-Forwarded-Proto": "https",
-                        "X-Forwarded-Host": "http://192.168.0.1",
-                    },
-                }),
-                trustedProxyHeaders: false,
-                redirectTo: "/",
-                expected: "/",
-            },
-            {
-                description: "with redirectTo parameter containing full URL with query",
-                request: new Request(signInURL),
-                trustedProxyHeaders: false,
-                redirectTo: "https://example.com/auth/signIn?next=123",
-                expected: "/auth/signIn?next=123",
-            },
-        ]
-
-        for (const { description, request, expected, redirectTo: redirectToParam, trustedProxyHeaders } of testCases) {
-            test(description, async () => {
-                const redirectTo = await createRedirectTo(request, redirectToParam, { trustedProxyHeaders } as GlobalContext)
-                expect(redirectTo).toBe(expected)
-            })
-        }
-    })
-
-    describe("invalid origin", () => {
-        const testCases = [
-            {
-                description: "different origins do not match",
-                request: new Request(signInURL, {
-                    headers: { Referer: "https://malicious.com/phishing" },
-                }),
-            },
-            {
-                description: "different protocols do not match",
-                request: new Request("https://www.example.com/auth/signIn/google", {
-                    headers: { Referer: "http://www.example.com/auth" },
-                }),
-            },
-            {
-                description: "different subdomains do not match",
-                request: new Request("https://example.com/auth/signIn/google", {
-                    headers: { Referer: "https://sub.example.com/auth" },
-                }),
-            },
-            {
-                description: "different ports do not match",
-                request: new Request("http://localhost:3000/auth/signIn/google", {
-                    headers: { Referer: "http://localhost:4000/auth" },
-                }),
-            },
-            {
-                description: "missing www in hosted URL",
-                request: new Request("https://www.example.com/auth/signIn/google", {
-                    headers: { Referer: "https://example.com/auth" },
-                }),
-            },
-
-            {
-                description: "invalid Referer URL (plain text)",
-                request: new Request(signInURL, {
-                    headers: { Referer: "not-a-valid-url" },
-                }),
-            },
-            {
-                description: "invalid protocol in Referer URL (unsupported protocol)",
-                request: new Request(signInURL, {
-                    headers: { Referer: "ftp://example.com/resource" },
-                }),
-            },
-            {
-                description: "encoded double slash in Referer path",
-                request: new Request(signInURL, {
-                    headers: { Referer: "https://%2f%2fexample.com" },
-                }),
-            },
-            {
-                description: "encoded double slash in Referer path",
-                request: new Request(signInURL, {
-                    headers: { Referer: "https://%2f%2fevil.com" },
-                }),
-            },
-            {
-                description: "invalid custom protocol in Referer URL",
-                request: new Request(signInURL, {
-                    headers: { Referer: "invalid://url" },
-                }),
-            },
-            {
-                description: "invalid path traversal using encoded segments",
-                request: new Request(signInURL, {
-                    headers: { Referer: "https://example.com/auth/%2e%2e/%2e%2e/admin" },
-                }),
-            },
-            {
-                description: "invalid path traversal using slashes",
-                request: new Request(signInURL, {
-                    headers: { Referer: "https://%2F%2Fexample.com" },
-                }),
-            },
-            {
-                description: "invalid Referer with spaces",
-                request: new Request(signInURL, {
-                    headers: { Referer: "https://example.com/auth path" },
-                }),
-            },
-            {
-                description: "invalid path traversal using encoded dots",
-                request: new Request(signInURL, {
-                    headers: { Referer: "https://example.com/%2e%2e/%2e%2e/" },
-                }),
-            },
-            {
-                description: "pathname contains invalid segments with backslashes",
-                request: new Request(signInURL, {
-                    headers: { Referer: "https://example.com/..\\admin" },
-                }),
-            },
-            {
-                description: "javascript URL injection in Referer query",
-                request: new Request("https://www.example.com/auth/signIn/google", {
-                    headers: { Referer: "https://www.example.com?redirectURL=javascript:alert(1)" },
-                }),
-            },
-            {
-                description: "javascript URL injection in Referer",
-                request: new Request(signInURL, {
-                    headers: { Referer: "javascript:alert(1)" },
-                }),
-            },
-            {
-                description: "with port in Referer header",
-                request: new Request(signInURL, {
-                    headers: { Referer: "https://example.com:3000" },
-                }),
-            },
-            {
-                description: "with IP address in Referer header",
-                request: new Request(signInURL, {
-                    headers: { Referer: "https://192.168.1.1/auth" },
-                }),
-            },
-            {
-                description: "with subdomain in Referer header",
-                request: new Request(signInURL, {
-                    headers: { Referer: "https://subdomain.example.com/auth" },
-                }),
-            },
-            {
-                description: "with .app domain in Referer header",
-                request: new Request(signInURL, {
-                    headers: { Referer: "https://example.app/auth" },
-                }),
-            },
-            {
-                description: "with .dev domain in Referer header",
-                request: new Request(signInURL, {
-                    headers: { Referer: "https://example.dev/auth" },
-                }),
-            },
-            {
-                description: "with redirectTo parameter containing full URL with different origin",
-                request: new Request(signInURL),
-                redirectTo: "https://malicious.com/auth/signIn",
-            },
-            {
-                description: "with redirectTo parameter containing invalid URL",
-                request: new Request(signInURL),
-                redirectTo: "ht!tp://invalid-url",
-            },
-        ]
-
-        for (const { description, request, redirectTo } of testCases) {
-            test(description, async () => {
-                expect(await createRedirectTo(request, redirectTo)).toEqual("/")
-            })
-        }
-    })
-
-    describe("with trustedOrigins config", () => {
-        test("accepts referer from trusted origin", async () => {
-            const request = new Request("https://example.com/auth/signIn/github", {
-                headers: { Referer: "https://example.com/dashboard" },
-            })
-            const redirectTo = await createRedirectTo(request, undefined, {
-                trustedOrigins: ["https://example.com", "https://admin.example.com"],
-            } as GlobalContext)
-            expect(redirectTo).toBe("/dashboard")
-        })
-
-        test("accepts referer matching wildcard pattern", async () => {
-            const request = new Request("https://app.example.com/auth/signIn/github", {
-                headers: { Referer: "https://app.example.com/dashboard" },
-            })
-            const redirectTo = await createRedirectTo(request, undefined, {
-                trustedOrigins: ["https://*.example.com"],
-            } as GlobalContext)
-            expect(redirectTo).toBe("/dashboard")
-        })
-
-        test("rejects referer not in trusted origins", async () => {
-            const request = new Request("https://example.com/auth/signIn/github", {
-                headers: { Referer: "https://malicious.com/phishing" },
-            })
-            const redirectTo = await createRedirectTo(request, undefined, {
-                trustedOrigins: ["https://example.com"],
-            } as GlobalContext)
-            expect(redirectTo).toBe("/")
-        })
-
-        test("accepts redirectTo from trusted origin", async () => {
-            const request = new Request("https://example.com/auth/signIn/github")
-            const redirectTo = await createRedirectTo(request, "https://example.com/dashboard", {} as GlobalContext)
-            expect(redirectTo).toBe("/dashboard")
-        })
-
-        test("with trusted origins that are not same origin", async () => {
-            const request = new Request("https://example.com/auth/signIn/github")
-            const redirectTo = await createRedirectTo(request, "https://api.example.com/redirect", {
-                trustedOrigins: ["https://api.example.com"],
-                trustedProxyHeaders: false,
-            } as GlobalContext)
-            expect(redirectTo).toBe("https://api.example.com/redirect")
-        })
-
-        test("misconfigurated", async () => {
-            const request = new Request("https://example.com/signIn/github")
-            const redirectTo = await createRedirectTo(request, "https://api.example.com/redirect", {
-                trustedOrigins: ["https://example.com"],
-                trustedProxyHeaders: false,
-            } as GlobalContext)
-            expect(redirectTo).toBe("/")
-        })
-
-        test("with wildcard pattern and subdomain and valid origin", async () => {
-            const request = new Request("https://app.example.com/auth/signIn/github")
-            const redirectTo = await createRedirectTo(request, "https://app.example.com/dashboard", {
-                trustedOrigins: ["https://*.example.com"],
-                trustedProxyHeaders: false,
-            } as GlobalContext)
-            expect(redirectTo).toBe("/dashboard")
-        })
-
-        test("with wildcard pattern and subdomain with invalid redirect value", async () => {
-            const request = new Request("https://app.example.com/auth/signIn/github")
-            const redirectTo = await createRedirectTo(request, "https://malicious.com/phishing", {
-                trustedOrigins: ["https://*.example.com"],
-                trustedProxyHeaders: false,
-            } as GlobalContext)
-            expect(redirectTo).toBe("/")
-        })
-
-        test("with wildcard pattern and subdomain with different subdomain", async () => {
-            const request = new Request("https://app.example.com/auth/signIn/github")
-            const redirectTo3 = await createRedirectTo(request, "https://sub.app.example.com/dashboard", {
-                trustedOrigins: ["https://*.example.com"],
-                trustedProxyHeaders: false,
-            } as GlobalContext)
-            expect(redirectTo3).toBe("/")
-        })
-
-        test("with wildcard pattern and subdomain with multiple subdomains", async () => {
-            const request = new Request("https://app.example.com/auth/signIn/github")
-            const redirectTo = await createRedirectTo(request, "https://evil.app.example.com/dashboard", {
-                trustedOrigins: ["https://*.example.com"],
-                trustedProxyHeaders: false,
-            } as GlobalContext)
-            expect(redirectTo).toBe("/")
-        })
-
-        test("with wildcard pattern with multiple subdomains", async () => {
-            const request = new Request("https://app.example.com/auth/signIn/github", {
-                headers: {
-                    "X-Forwarded-Proto": "https",
-                    "X-Forwarded-Host": "evil.app.example.com",
-                },
-            })
-            const redirectTo = await createRedirectTo(request, "https://evil.app.example.com/dashboard", {
-                trustedOrigins: ["https://*.example.com"],
-                trustedProxyHeaders: true,
-            } as GlobalContext)
-            expect(redirectTo).toBe("/")
-        })
-    })
-
-    describe("SECURITY", () => {
-        // This test demonstrates a potential security risk if trustedProxyHeaders is enabled in an
-        // insecure environment. It shows that if an attacker can control the X-Forwarded-Host header,
-        // they could potentially bypass the trusted origins check and redirect users to a malicious site.
-        test("with wildcard pattern with dangerous origin from trustedProxyHeaders", async () => {
-            const request = new Request("https://app.example.com/auth/signIn/github", {
-                headers: {
-                    "X-Forwarded-Proto": "https",
-                    "X-Forwarded-Host": "evil.example.com",
-                },
-            })
-            const redirectTo = await createRedirectTo(request, "https://evil.example.com/dashboard", {
-                trustedOrigins: ["https://*.example.com"],
-                trustedProxyHeaders: true,
-            } as GlobalContext)
-            expect(redirectTo).toBe("https://evil.example.com/dashboard")
-        })
-
-        test("with wildcard pattern with trustedProxyHeaders", async () => {
-            const request = new Request("https://app.example.com/auth/signIn/github", {
-                headers: {
-                    "X-Forwarded-Proto": "https",
-                    "X-Forwarded-Host": "app.example.com",
-                },
-            })
-            const redirectTo = await createRedirectTo(request, "https://evil.example.com/dashboard", {
-                trustedOrigins: ["https://*.example.com"],
-                trustedProxyHeaders: true,
-            } as GlobalContext)
-            expect(redirectTo).toBe("https://evil.example.com/dashboard")
-        })
-    })
-})
-
-describe("getOriginURL", () => {
-    const testCases = [
-        {
-            description: "with standard URL",
-            request: new Request("https://example.com/auth/signIn/github"),
-            trustedProxyHeaders: false,
-            trustedOrigins: [],
-            expected: "https://example.com",
-        },
-        {
-            description: "with localhost URL",
-            request: new Request("http://localhost:3000/auth/signIn/github"),
-            trustedProxyHeaders: false,
-            trustedOrigins: [],
-            expected: "http://localhost:3000",
-        },
-        {
-            description: "with IP address URL",
-            request: new Request("http://192.168.0.1/auth/signIn/github"),
-            trustedProxyHeaders: false,
-            trustedOrigins: [],
-            expected: "http://192.168.0.1",
-        },
-        {
-            description: "without trusted proxy headers",
-            request: new Request("http://localhost:3000/auth/signIn/github", {
-                headers: {
-                    "X-Forwarded-Proto": "https",
-                    "X-Forwarded-Host": "example.com",
-                },
-            }),
-            trustedProxyHeaders: false,
-            trustedOrigins: [],
-            expected: "http://localhost:3000",
-        },
-        {
-            description: "with trusted proxy headers",
-            request: new Request("http://localhost:3000/auth/signIn/github", {
-                headers: {
-                    "X-Forwarded-Proto": "https",
-                    "X-Forwarded-Host": "example.com",
-                },
-            }),
-            trustedProxyHeaders: true,
-            trustedOrigins: ["https://example.com"],
-            expected: "https://example.com",
-        },
-        {
-            description: "with localhost and trusted proxy headers",
-            request: new Request("http://localhost:3000/auth/signIn/github", {
-                headers: {
-                    "X-Forwarded-Proto": "http",
-                    "X-Forwarded-Host": "192.168.0.1",
-                },
-            }),
-            trustedProxyHeaders: true,
-            trustedOrigins: ["http://192.168.0.1"],
-            expected: "http://192.168.0.1",
-        },
-        {
-            description: "priority of Forwarded header over X-Forwarded headers",
-            request: new Request("http://localhost:3000/auth/signIn/github", {
-                headers: {
-                    Forwarded: "proto=https;host=app.com",
-                    "X-Forwarded-Proto": "http",
-                    "X-Forwarded-Host": "malicious.com",
-                },
-            }),
-            trustedProxyHeaders: true,
-            trustedOrigins: ["https://app.com"],
-            expected: "https://app.com",
-        },
-    ]
-
-    for (const { description, request, trustedProxyHeaders, trustedOrigins, expected } of testCases) {
-        test(description, async () => {
-            const originURL = await getOriginURL(request, { trustedProxyHeaders, trustedOrigins } as GlobalContext)
-            expect(originURL).toBe(expected)
-        })
-    }
-
-    describe("Invalid origins", () => {
-        const testCases = [
-            {
-                description: "origin not set in trustedOrigins",
-                request: new Request("https://example.com/signIn/github", {
-                    headers: {
-                        "X-Forwarded-Proto": "http",
-                        "X-Forwarded-Host": "localhost:3000",
-                    },
-                }),
-                trustedOrigins: ["https://example.com"],
-                trustedProxyHeaders: true,
-            },
-            {
-                description: "trustedOrigins empty",
-                request: new Request("https://example.com/signIn/github", {
-                    headers: {
-                        "X-Forwarded-Proto": "http",
-                        "X-Forwarded-Host": "localhost:3000",
-                    },
-                }),
-                trustedOrigins: [],
-                trustedProxyHeaders: true,
-            },
-            {
-                description: "with trusted proxy headers and without trusted origins",
-                request: new Request("http://localhost:3000/auth/signIn/github", {
-                    headers: {
-                        "X-Forwarded-Proto": "https",
-                        "X-Forwarded-Host": "malicious.com",
-                    },
-                }),
-                trustedOrigins: [],
-                trustedProxyHeaders: true,
-            },
-        ]
-
-        for (const { description, request, trustedOrigins, trustedProxyHeaders } of testCases) {
-            test(description, async () => {
-                await expect(getOriginURL(request, { trustedOrigins, trustedProxyHeaders } as GlobalContext)).rejects.toThrow()
-            })
-        }
-    })
-})
-
-describe("getBaseURL", () => {
-    const testCases = [
-        {
-            description: "baseURL from request URL",
-            request: new Request("https://example.com/auth/signIn/github"),
-            trustedProxyHeaders: false,
-            expected: "https://example.com",
-        },
-        {
-            description: "baseURL from request URL with trustedProxyHeaders",
-            request: new Request("https://example.com/auth/signIn/github", {
-                headers: {
-                    "X-Forwarded-Proto": "https",
-                    "X-Forwarded-Host": "example.com",
-                },
-            }),
-            trustedProxyHeaders: true,
-            expected: "https://example.com",
-        },
-        {
-            description: "baseURL from request URL with trusted proxy headers and localhost",
-            request: new Request("http://localhost:3000/auth/signIn/github", {
-                headers: {
-                    "X-Forwarded-Proto": "http",
-                    "X-Forwarded-Host": "localhost:4000",
-                },
-            }),
-            trustedProxyHeaders: true,
-            expected: "http://localhost:4000",
-        },
-        {
-            description: "baseURL from request object with trustedProxyHeaders and priority headers",
-            request: new Request("https://example.com/auth/signIn/github", {
-                headers: {
-                    Forwarded: "proto=https;host=example.com",
-                    "X-Forwarded-Proto": "http",
-                    "X-Forwarded-Host": "malicious.com",
-                },
-            }),
-            trustedProxyHeaders: true,
-            expected: "https://example.com",
-        },
-    ]
-
-    for (const { description, request, trustedProxyHeaders, expected } of testCases) {
-        test(description, async () => {
-            const baseURL = await getBaseURL({ request, ctx: { trustedProxyHeaders } as GlobalContext })
-            expect(baseURL).toBe(expected)
-        })
-    }
-
-    test("with BASE_URL environment variable", async () => {
-        vi.stubEnv("BASE_URL", "https://example.com")
-        const baseURL = await getBaseURL({})
-        expect(baseURL).toBe("https://example.com")
-    })
-
-    test("with baseURL in context", async () => {
-        const baseURL = await getBaseURL({ ctx: { baseURL: "https://example.com" } as GlobalContext })
-        expect(baseURL).toBe("https://example.com")
-    })
-})
