+++ packages/core/test/helpers/secure.test.ts
